Moved postcode list from utils to postcode pages directory
